Extract methods into pure functions
